refactor: 기존의 컨트롤러 테스트들을 목적에 따라 인수 테스트를 의미하는 /acceptance 로 이동 및 클래스명 수정
